+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## v1.5.3 - Logging improvements & new Settings
+
+### Added
+
+- There's now an option to save server logs to a `.log` file ([Issue #65](https://github.com/Yooli8537/Editor/issues/65)).
+- Just like for images, there's an option to clear out all log files (except for the one that's currently in use).
+- Documented some API Endpoints, not all because it kinda takes forever.
+- Exports no longer have to be confirmed, can be changed in settings ([Issue #66](https://github.com/Yooli8537/Editor/issues/66)).
+- The Settings page now has URLs ([Issue #61](https://github.com/Yooli8537/Editor/issues/61)).
+
+### Removed
+
+- Removed highlight from error modal because it's ugly.
+
+### Fixed
+
+- Server logs now show the correct time + show the current date ([Issue #51](https://github.com/Yooli8537/Editor/issues/51)).
+
 ## v1.5.2 - Minor improvements & Bug fixes
 
 ### Added